Abstracted `StakeInputArea`
